生成service 目前 完成service的生成 将Mapper、service方法返回值由Integer改成Long

+++ easyjava-demo/src/main/java/com/easyjava/mappers/ProductInfoMapper.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.query.Param;
 /**
  * @Description: 商品信息Mapper对象
  * @author: 小阙
- * @date: 2023/05/19
+ * @date: 2023/05/20
  */
 public interface ProductInfoMapper<T, P> extends BaseMapper {
 	/**
@@ -16,12 +16,12 @@ public interface ProductInfoMapper<T, P> extends BaseMapper {
 	/**
 	 * 根据Id更新
 	 */
-	Integer updateById(@Param("bean") T t, @Param("id") Integer id);
+	Long updateById(@Param("bean") T t, @Param("id") Integer id);
 
 	/**
 	 * 根据Id删除
 	 */
-	Integer deleteById(@Param("id") Integer id);
+	Long deleteById(@Param("id") Integer id);
 
 	/**
 	 * 根据Code查询
@@ -31,12 +31,12 @@ public interface ProductInfoMapper<T, P> extends BaseMapper {
 	/**
 	 * 根据Code更新
 	 */
-	Integer updateByCode(@Param("bean") T t, @Param("code") String code);
+	Long updateByCode(@Param("bean") T t, @Param("code") String code);
 
 	/**
 	 * 根据Code删除
 	 */
-	Integer deleteByCode(@Param("code") String code);
+	Long deleteByCode(@Param("code") String code);
 
 	/**
 	 * 根据SkuTypeAndColorType查询
@@ -46,12 +46,12 @@ public interface ProductInfoMapper<T, P> extends BaseMapper {
 	/**
 	 * 根据SkuTypeAndColorType更新
 	 */
-	Integer updateBySkuTypeAndColorType(@Param("bean") T t, @Param("skuType") Integer skuType, @Param("colorType") String colorType);
+	Long updateBySkuTypeAndColorType(@Param("bean") T t, @Param("skuType") Integer skuType, @Param("colorType") String colorType);
 
 	/**
 	 * 根据SkuTypeAndColorType删除
 	 */
-	Integer deleteBySkuTypeAndColorType(@Param("skuType") Integer skuType, @Param("colorType") String colorType);
+	Long deleteBySkuTypeAndColorType(@Param("skuType") Integer skuType, @Param("colorType") String colorType);
 
 
 }
