+++ easyjava-demo/src/main/java/com/easyjava/mappers/UserMapper.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.query.Param;
 /**
  * @Description: 用户表Mapper对象
  * @author: 小阙
- * @date: 2023/05/19
+ * @date: 2023/05/20
  */
 public interface UserMapper<T, P> extends BaseMapper {
 	/**
@@ -16,12 +16,12 @@ public interface UserMapper<T, P> extends BaseMapper {
 	/**
 	 * 根据Id更新
 	 */
-	Integer updateById(@Param("bean") T t, @Param("id") Integer id);
+	Long updateById(@Param("bean") T t, @Param("id") Integer id);
 
 	/**
 	 * 根据Id删除
 	 */
-	Integer deleteById(@Param("id") Integer id);
+	Long deleteById(@Param("id") Integer id);
 
 
 }
